requirements.txt
pandas
psycopg2-binary
python-decouple
tkcalendar